split mock visitors
